build fix for format string warning
git-svn-id: https://svn.macosforge.org/repository/darwinbuild/trunk@867 10a61168-4876-4dac-953b-31e694342555

+++ darwinxref/plugins/loadDeps.c
@@ -97,7 +97,7 @@ int loadDeps(const char* build, const char* project, const char *root) {
 			struct stat sb;
 			int typesize = (intptr_t)tab - (intptr_t)line;
 			asprintf(&type, "%.*s", typesize, line);
-			asprintf(&file, "%.*s", size - typesize - 1, tab+1);
+			asprintf(&file, "%.*s", (int)size - typesize - 1, tab+1);
 			if (strcmp(type, "open") == 0) {
 				free(type);
 				if (has_suffix(file, ".h")) {
